+++ CHANGELOG.md
@@ -6,3 +6,12 @@ Update [mpv-remote-node](https://github.com/husudosu/mpv-remote-node) to 1.0.3!
 - Metadata appears on info panel
 - Some tweaking
 - Unused dependecies removed from project
+
+## 1.0.5
+
+Very important to update mpv-remote-node to 1.0.5!
+
+- Upgraded to Ionic v6,
+- Support for multiple servers,
+- Multiple server handling
+- Filebrowser speed improvements with larger directories/collections